split data_structures module
