+++ internal/test_cases/zadd_test_case.go
@@ -3,7 +3,7 @@ package test_cases
 import (
 	"strconv"
 
-	"github.com/codecrafters-io/redis-tester/internal/data_structures"
+	"github.com/codecrafters-io/redis-tester/internal/data_structures/sorted_set"
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_connection"
 	"github.com/codecrafters-io/redis-tester/internal/resp_assertions"
 	"github.com/codecrafters-io/tester-utils/logger"
@@ -11,7 +11,7 @@ import (
 
 type ZaddTestCase struct {
 	Key                       string
-	Member                    data_structures.SortedSetMember
+	Member                    sorted_set.SortedSetMember
 	ExpectedAddedMembersCount int
 }
