+++ internal/test_zset_zadd2.go
@@ -1,7 +1,7 @@
 package internal
 
 import (
-	"github.com/codecrafters-io/redis-tester/internal/data_structures"
+	"github.com/codecrafters-io/redis-tester/internal/data_structures/sorted_set"
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_connection"
 	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"github.com/codecrafters-io/redis-tester/internal/test_cases"
@@ -24,7 +24,7 @@ func testZsetZadd2(stageHarness *test_case_harness.TestCaseHarness) error {
 	defer client.Close()
 
 	zsetKey := testerutils_random.RandomWord()
-	sortedSet := data_structures.GenerateSortedSetWithRandomMembers(data_structures.SortedSetMemberGenerationOption{
+	sortedSet := sorted_set.GenerateSortedSetWithRandomMembers(sorted_set.SortedSetMemberGenerationOption{
 		Count: testerutils_random.RandomInt(2, 4),
 	})
 	members := sortedSet.GetMembers()
@@ -43,10 +43,10 @@ func testZsetZadd2(stageHarness *test_case_harness.TestCaseHarness) error {
 
 	// Test by updating an existing member
 	memberToUpdate := members[testerutils_random.RandomInt(0, sortedSet.Size())]
-	newScore := data_structures.GetRandomSortedSetScore()
+	newScore := sorted_set.GetRandomSortedSetScore()
 	zaddTestCase := test_cases.ZaddTestCase{
 		Key: zsetKey,
-		Member: data_structures.SortedSetMember{
+		Member: sorted_set.SortedSetMember{
 			Name:  memberToUpdate.Name,
 			Score: newScore,
 		},
