+++ internal/test_zset_zrange_neg_index.go
@@ -1,7 +1,7 @@
 package internal
 
 import (
-	"github.com/codecrafters-io/redis-tester/internal/data_structures"
+	"github.com/codecrafters-io/redis-tester/internal/data_structures/sorted_set"
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_connection"
 	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"github.com/codecrafters-io/redis-tester/internal/test_cases"
@@ -24,7 +24,7 @@ func testZsetZrangeNegIndex(stageHarness *test_case_harness.TestCaseHarness) err
 	defer client.Close()
 
 	zsetKey := testerutils_random.RandomWord()
-	sortedSet := data_structures.GenerateSortedSetWithRandomMembers(data_structures.SortedSetMemberGenerationOption{
+	sortedSet := sorted_set.GenerateSortedSetWithRandomMembers(sorted_set.SortedSetMemberGenerationOption{
 		Count:          testerutils_random.RandomInt(3, 5),
 		SameScoreCount: 2,
 	})
