+++ internal/test_zset_zscore.go
@@ -1,7 +1,7 @@
 package internal
 
 import (
-	"github.com/codecrafters-io/redis-tester/internal/data_structures"
+	"github.com/codecrafters-io/redis-tester/internal/data_structures/sorted_set"
 	"github.com/codecrafters-io/redis-tester/internal/instrumented_resp_connection"
 	"github.com/codecrafters-io/redis-tester/internal/redis_executable"
 	"github.com/codecrafters-io/redis-tester/internal/resp_assertions"
@@ -25,7 +25,7 @@ func testZsetZscore(stageHarness *test_case_harness.TestCaseHarness) error {
 	defer client.Close()
 
 	zsetKey := testerutils_random.RandomWord()
-	sortedSet := data_structures.GenerateSortedSetWithRandomMembers(data_structures.SortedSetMemberGenerationOption{
+	sortedSet := sorted_set.GenerateSortedSetWithRandomMembers(sorted_set.SortedSetMemberGenerationOption{
 		Count:          testerutils_random.RandomInt(3, 5),
 		SameScoreCount: 2,
 	})
@@ -57,10 +57,10 @@ func testZsetZscore(stageHarness *test_case_harness.TestCaseHarness) error {
 	}
 
 	// Update an existing member
-	newScore := data_structures.GetRandomSortedSetScore()
+	newScore := sorted_set.GetRandomSortedSetScore()
 	zaddTestCase := test_cases.ZaddTestCase{
 		Key: zsetKey,
-		Member: data_structures.SortedSetMember{
+		Member: sorted_set.SortedSetMember{
 			Name:  memberToTest.Name,
 			Score: newScore,
 		},
